+++ packages/react-api/src/api/common.js
@@ -1,7 +1,7 @@
 /**
  * Return more descriptive error from API
  */
-export function dealWithApiError({ credentials, response, data }) {
+export function dealWithApiError({ response, data }) {
   switch (response.status) {
     case 401:
       throw new Error('Unauthorized access. Invalid credentials');
@@ -10,4 +10,41 @@ export function dealWithApiError({ credentials, response, data }) {
     default:
       throw new Error(`${JSON.stringify(data?.hint || data.error?.[0])}`);
   }
+}
+
+export function assert(condition, message) {
+  if (!condition) {
+    throw new Error(message);
+  }
+}
+
+export function checkCredentials(credentials) {
+  if (!credentials || !credentials.apiBaseUrl || !credentials.accessToken) {
+    throw new Error('Missing or bad credentials provided');
+  }
+}
+
+export async function makeCall({ url, credentials, opts }) {
+  let response;
+  let data;
+  try {
+    response = await fetch(url.toString(), {
+      headers: {
+        Authorization: `Bearer ${credentials.accessToken}`
+      },
+      signal: opts?.abortController?.signal,
+      ...opts?.otherOptions
+    });
+    data = await response.json();
+  } catch (error) {
+    if (error.name === 'AbortError') throw error;
+
+    throw new Error(`Failed request: ${error}`);
+  }
+
+  if (!response.ok) {
+    dealWithApiError({ response, data });
+  }
+
+  return data;
 }