+++ packages/react-api/src/api/lds.js
@@ -1,4 +1,4 @@
-import { dealWithApiError } from './common';
+import { checkCredentials, makeCall } from './common';
 
 /**
  * Execute a LDS geocoding service geocode request.
@@ -13,9 +13,8 @@ import { dealWithApiError } from './common';
  * @param { Object= } props.opts - Additional options for the HTTP request
  */
 export async function ldsGeocode({ credentials, address, country, limit, opts }) {
-  if (!credentials || !credentials.apiBaseUrl || !credentials.accessToken) {
-    throw new Error('ldsGeocode: Missing or bad credentials provided');
-  }
+  checkCredentials(credentials);
+
   if (!address) {
     throw new Error('ldsGeocode: No address provided');
   }
@@ -29,29 +28,10 @@ export async function ldsGeocode({ credentials, address, country, limit, opts })
     url.searchParams.set('limit', String(limit));
   }
 
-  const { abortController, ...otherOptions } = opts || {};
+  let data = await makeCall({ url, credentials, opts });
 
-  let response;
-  let data;
-  try {
-    response = await fetch(url.toString(), {
-      headers: {
-        Authorization: `Bearer ${credentials.accessToken}`
-      },
-      signal: abortController?.signal,
-      ...otherOptions
-    });
-    data = await response.json();
-  } catch (error) {
-    if (error.name === 'AbortError') throw error;
-
-    throw new Error(`Failed to connect to LDS API: ${error}`);
-  }
   if (Array.isArray(data)) {
     data = data[0];
-  }
-  if (!response.ok) {
-    dealWithApiError({ credentials, response, data });
   }
 
   return data.value;