+++ packages/react-api/src/index.d.ts
@@ -1,5 +1,7 @@
 export { executeSQL } from './api/SQL';
 export { ldsGeocode } from './api/lds';
+export { getStats as _getStats } from './api/stats';
+export { getTileJson as _getTileJson } from './api/tilejson';
 
 export { default as useCartoLayerProps } from './hooks/useCartoLayerProps';