+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Not released
 
+- Implement Stats API fn [#404](https://github.com/CartoDB/carto-react/pull/404)
 - Fix map filtering with CategoryWidget using boolean values [#411](https://github.com/CartoDB/carto-react/pull/411)
-
 
 ## 1.3